// src/types/side-bar.type.ts
export enum SIDEBAR_TITLE {
  HOME = "Home",
  SEARCH = "Search",
  EXPLORE = "Explore",
  NOTIFICATIONS = "Notifications",
  MESSAGES = "Messages",
  LISTS = "Lists",
  PROFILE = "Profile",
  MORE = "More",
  CREATE_POST = "Create Post",
}

export enum EXPANDED_CONTENT_TYPE {
  SEARCH = "Search",
  NOTIFICATIONS = "Notifications",
  MESSAGES = "Messages",
  PROFILE = "Profile",
}